ROTI: WIP checkpoint from aborted agent run (2026-04-20)
Prior core-engineer agent (a7fe9ca547748e383) completed implementation
but aborted before /pr-submit; work preserved here for pickup.

Net: +48 / -151 LOC across 4 files (proposal projected -80 to -90, so
likely complete). No quality-gate run, no Tier-2 run, no PR opened.
Resuming agent should rebase on origin/main, run /quality-gate +
/test --tier2 (symmetry is Tier-2 trigger), and /pr-submit.

+++ pwdft/pwdft-core/src/symmetry/density/mod.rs
@@ -37,10 +37,7 @@ pub fn check_grid_compatibility(dims: [usize; 3], symmetry: &SymmetryInfo) -> bo
         let r = &op.rotation;
         for i in 0..3 {
             for j in 0..3 {
-                // `r[i][j]` is `i8` (TYPE-A); widen to `i32` for the product
-                // with the FFT dim. Compatibility test is `i32`-sized
-                // regardless of storage width.
-                if (i32::from(r[i][j]) * ns[j]) % ns[i] != 0 {
+                if (r[i][j] * ns[j]) % ns[i] != 0 {
                     return false;
                 }
             }
